Update parser.py
Attempted to remove warning messages but did not succeed. Might work on other computers so left in code. Reverted from onnixt5 back to original because onnxt5 uses tqdm module that displays progress bars.. This slows the model down a lot and no way to remove it. Although onnxt5 might be quicker, i cannot test it because of this internal library.